pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astro-neo"
version = "0.1.0"
description = "ASTRO-NEO orchestration services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi~=0.115",
    "uvicorn[standard]~=0.30",
    "pydantic-settings~=2.4",
    "pyyaml~=6.0",
    "httpx~=0.27",
    "jinja2>=3.1",
    "python-multipart~=0.0.9",
    "numpy~=2.1",
    "astropy~=6.1",
    "astroplan~=0.9",
    "sqlmodel~=0.0.16",
    "alembic~=1.13",
    "psycopg[binary,pool]~=3.1",
    "prometheus-client~=0.20",
    "photutils~=1.12",
    "Pillow~=10.4",
    "scipy~=1.11",
    "lxml~=5.1",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.0",
    "httpx~=0.27",
]

[tool.setuptools.packages.find]
where = ["app"]
